+++ packages/routed/lib/src/router/router_group_builder.dart
@@ -26,4 +26,10 @@ class RouterGroupBuilder {
     _router.groupName = groupName;
     return this;
   }
+
+  /// Excludes middleware from this group.
+  RouterGroupBuilder withoutMiddleware(Iterable<Object> middlewares) {
+    _router.excludeMiddlewares(middlewares);
+    return this;
+  }
 }